fix(lowering): monotonic_millis returns int, not double (#634)
`let t = monotonic_millis()` inferred `t :: double` because the runtime
helper registry in `compiler/src/llvm/runtime_helpers.sfn` declared
both `monotonic_millis` and `runtime_monotonic_millis_fn` with
`return_type: "double"`. Every downstream `int`-typed call boundary
(`_ms_since(t)`, `t1 - t0`, etc.) then emitted an `ABI primitive
mismatch` warning at LLVM lowering — six warnings per
`compile_to_llvm_with_module_timed` invocation, plus the same shape
at the `cli_commands.sfn` test-runner timing sites.

Root cause was a stale C ABI: `sailfin_runtime_monotonic_millis`
returned `double` even though the prelude declared the wrapper as
`fn monotonic_millis() -> int ![clock]`. The fix moves the C ABI to
`int64_t` and points the registry at `"i64"`, so the inference matches
the user-facing declaration end-to-end.

Self-host consideration: the seed (v0.5.10-alpha.19) still emits IR
that calls `double @sailfin_runtime_monotonic_millis()` against the
new `int64_t` C symbol while building the first-pass binary. The ABI
mismatch only affects timing values (the seed's IR reads %xmm0 while
the C function writes %rax), and timings are purely cosmetic — no
control flow depends on them. The first-pass binary then compiles
clean second-pass IR because its embedded registry has the new
`"i64"` shape.

Closes #634

+++ runtime/native/include/sailfin_runtime.h
@@ -96,8 +96,13 @@ extern "C"
     void sfn_sleep(double milliseconds);
 
     // Monotonic clock for profiling/timing.
-    // Returns an integer millisecond count as a `double` for stage2 ABI.
-    double sailfin_runtime_monotonic_millis(void);
+    // Returns an integer millisecond count as `int64_t` (Sailfin `int`).
+    // Pre-#634: returned `double` for legacy stage2 ABI; the registry in
+    // `compiler/src/llvm/runtime_helpers.sfn` mirrored that and forced
+    // every `let t = monotonic_millis()` to infer `double`, producing
+    // spurious ABI primitive mismatch warnings at every `int`-typed
+    // call site downstream.
+    int64_t sailfin_runtime_monotonic_millis(void);
 
     // ---- String helpers ----
